chore:  Add initial boilerplate setup
Configured:
- Webpack
- Babel
- Jest
- Enzyme
- Eslint
- Prettier
- Husky
- Project scaffolding

Add the initial setup for the project to allow to use
Webpack as bundle generator, babel as transpiler,
sass as style preprocessor, jest and enzyme as testing suite
and  configure tools like eslint, prettier and husky
for enforce the development best practices.
Also provide an initial scaffolding for a react.js project.